created endpoints
auth endpoint
get /auth with token param returns the userid
delete /auth with token param checks if token is valid and deletes it
post /auth with token param checks if token is valid, and ties the user belonging to a token to a new one.
post /auth with body of username and password is used to login the user and return a new token if valid.

register endpoint
post /register registers email and password if validation is passes